+++ tests/deploy/test_nemo_deploy.py
@@ -24,7 +24,8 @@ import pytest
 import torch
 from tqdm import tqdm
 
-from nemo.deploy import DeployPyTriton, NemoQueryLLM
+from nemo.deploy import DeployPyTriton
+from nemo.deploy.nlp import NemoQueryLLM
 from nemo.export import TensorRTLLM
 from tests.infer_data_path import download_nemo_checkpoint, get_infer_test_data
